Add UserCompany, it's requirements and authentication
User must relate to many Companies, and Company must have many Users.
Also user must authenticate in order to company requests return 200.

It add model, table, tests and configurations required to make this
relation.
After company is created, it will create a record in the UserCompany
with the current_user and the company just created.